+++ app/models/category.rb
@@ -3,6 +3,8 @@ class Category < ApplicationRecord
 
   has_many :transactions, dependent: :destroy
   has_many :budgets, dependent: :destroy
+  has_many :recurring_incomes, dependent: :destroy
+  has_many :recurring_bills, dependent: :destroy
 
   validates :name, presence: true
   validates :kind, inclusion: { in: KINDS }